rag-backend/requirements.txt
fastapi
uvicorn
python-dotenv
llama-index
openai
faiss-cpu
numpy
Stemmer
